Fix double call for doFilter

+++ src/main/java/com/example/usermanager/repository/CustomerRepository.java
@@ -4,11 +4,9 @@ import com.example.usermanager.domain.Customer;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
 public interface CustomerRepository extends MongoRepository<Customer, Long> {
 
     Customer findByLogin(String login);
-    Customer findByCustomerId(BigInteger customerId);
+    Customer findByCustomerId(Long customerId);
 }
